refactor: cupons

+++ apps/api/src/modules/cupons/controller.ts
@@ -1,22 +1,12 @@
 import { eq } from "drizzle-orm"
 import type { Request, Response } from "express"
-import { z } from "zod"
 import { db } from "~/db/db"
 import { cupons } from "~/db/schema"
 import { paramsSchema } from "~/shared/schemas"
-
-const cadastrarCupomSchema = z.object({
-  codigo: z.string().toUpperCase(),
-  valor: z.number().int().positive(),
-})
-
-const atualizarCupomSchema = z.object({
-  codigo: z.string().toUpperCase().optional(),
-  valor: z.number().int().positive().optional(),
-})
+import { createCupomSchema, updateCupomSchema } from "./dtos"
 
 async function cadastrarCupom(req: Request, res: Response) {
-  const data = cadastrarCupomSchema.parse(req.body)
+  const data = createCupomSchema.parse(req.body)
 
   await db.insert(cupons).values(data)
 
@@ -24,7 +14,7 @@ async function cadastrarCupom(req: Request, res: Response) {
   return
 }
 
-async function listarCupons(req: Request, res: Response) {
+async function listarCupons(_: Request, res: Response) {
   const cupons = await db.query.cupons.findMany()
 
   res.status(200).json(cupons)
@@ -38,6 +28,11 @@ async function mostrarCupom(req: Request, res: Response) {
     where: (cupons, { eq }) => eq(cupons.id, id),
   })
 
+  if (!cupom) {
+    res.status(401).json({ message: "Cupom não encontrado" })
+    return
+  }
+
   res.status(200).json(cupom)
   return
 }
@@ -45,7 +40,16 @@ async function mostrarCupom(req: Request, res: Response) {
 async function atualizarCupom(req: Request, res: Response) {
   const { id } = paramsSchema.parse(req.params)
 
-  const data = atualizarCupomSchema.parse(req.body)
+  const cupom = await db.query.cupons.findFirst({
+    where: (cupons, { eq }) => eq(cupons.id, id),
+  })
+
+  if (!cupom) {
+    res.status(401).json({ message: "Cupom não encontrado" })
+    return
+  }
+
+  const data = updateCupomSchema.parse(req.body)
 
   await db.update(cupons).set(data).where(eq(cupons.id, id))
 
@@ -55,6 +59,15 @@ async function atualizarCupom(req: Request, res: Response) {
 
 async function deletarCupom(req: Request, res: Response) {
   const { id } = paramsSchema.parse(req.params)
+
+  const cupom = await db.query.cupons.findFirst({
+    where: (cupons, { eq }) => eq(cupons.id, id),
+  })
+
+  if (!cupom) {
+    res.status(401).json({ message: "Cupom não encontrado" })
+    return
+  }
 
   await db.delete(cupons).where(eq(cupons.id, id))
